apartments_service/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers>=4.3.1
tensorflow>=2.15.0
scikit-learn>=1.3.2
numpy>=1.24.0
joblib>=1.3.2
Pillow>=10.0.0
psycopg2-binary
pandas>=2.1.0